backend/discord-mcp-server/requirements.txt
mcp
discord.py>=2.3.0
python-dotenv
pydantic
asyncio